backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary
redis
pydantic_settings
pytest
flake8
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart